refactor: 유저 자동완성 응답 클래스 이름 변경
